split app.ts into modules using es6 modules and reconfigured tsconfig.js to compile with modules
